Move populating users to a dedicated function
